Extract public CLI into its own module
